proyectofinal/requirements.txt
arabic-reshaper==2.1.1
asgiref==3.3.0
astroid==2.4.2
autopep8==1.4.4
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.4
dj-database-url==0.5.0
Django==2.0
django-versatileimagefield==2.0
django-widget-tweaks==1.4.1
djangorestframework==3.7.3
future==0.18.2
gunicorn==20.0.4
html5lib==1.1
idna==2.8
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==8.0.1
pycodestyle==2.6.0
pylint==2.6.0
pyparsing==2.4.7
PyPDF2==1.26.0
python-bidi==0.4.2
python-magic==0.4.18
pytoml==0.1.21
pytz==2017.3
reportlab==3.5.55
requests==2.22.0
retrying==1.3.3
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
toml==0.10.2
urllib3==1.25.11
webencodings==0.5.1
whitenoise==5.2.0
wrapt==1.12.1
xhtml2pdf==0.2.5
youtube-dl==2020.11.26
